requirements.txt
streamlit
cohere
faiss-cpu
langchain
PyPDF2
numpy
python-dotenv